rust: drm: ioctl: Add DRM ioctl abstraction
DRM drivers need to be able to declare which driver-specific ioctls they
support. This abstraction adds the required types and a helper macro to
generate the ioctl definition inside the DRM driver.

Note that this macro is not usable until further bits of the
abstraction are in place (but it will not fail to compile on its own, if
not called).

v2: Rebased on top of rust-next (6.6-rc1)
    each commit needs polish and review.

updated code

[rust/kernel/drm/ioctl.rs]
+use crate::uapi::DRM_IOCTL_BASE;

-const BASE: u32 = bindings::DRM_IOCTL_BASE as u32;
+const BASE: u32 = DRM_IOCTL_BASE as u32;

Signed-off-by: Asahi Lina <lina@asahilina.net>
Signed-off-by: Gwan-gyeong Mun <gwan-gyeong.mun@intel.com>