rust: platform: add direct registration for a platform device
Allows drivers to directly allocate and register a device. Although this
style of driver is discouraged, it is useful for virtual devices that
cannot be probed. Also, handles the deallocation of a registered device
through the Drop trait.

G.G.: Rebase and replace `from_kernel_err_ptr` to `from_err_ptr`

Signed-off-by: Maíra Canal <mcanal@igalia.com>
Signed-off-by: Gwan-gyeong Mun <gwan-gyeong.mun@intel.com>