drm/rustgem: implement a Rust version of VGEM
The VGEM driver is a non-hardware backed GEM service and it is currently
implemented in C on the DRM subsystem. This patch introduces a new
version of the VGEM driver written in Rust. This driver provides the
same functionalities and uses the same UAPI of the original VGEM driver.

G.G.: Rebased on top of Linux 6.6
      Fix to use updated pin,xarray,uapi binding etc.

Signed-off-by: Maíra Canal <mcanal@igalia.com>
Signed-off-by: Gwan-gyeong Mun <gwan-gyeong.mun@intel.com>